greengrass-components/applications/docker_certificate_rotator/certificate_rotator_app/requirements.txt
awsiotsdk==1.21.0
cryptography==41.0.7